phase_2/core/requirements.txt
streamlit
langchain
langchain-community
langchain-core
openai
faiss-cpu
tiktoken
pypdf
python-docx
numpy
pandas
scikit-learn
python-dotenv